requirements.txt
construct
lief
pycryptodome
